Refactor to separate the renderer from the main program.
