+++ Src/DXRenderer.cpp
@@ -3,17 +3,15 @@
 #include <Windows.h>
 
 // Renderer
+#include "pch.h"
 #include "MainWindow.h"
-#include "USDReader.h"
+#include "USDScene.h"
 
 // Force Nvidia GPU - e.g in laptop cases.
 extern "C" { __declspec(dllexport) unsigned long NvOptimusEnablement = 0x00000001; }
 
 int WINAPI WinMain( _In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance, _In_ LPSTR lpCmdLine, _In_ int nCmdShow)
 {
-    USDReader Scene = USDReader();
-    Scene.LoadExampleCube();
-    
     MainWindow App = MainWindow(hInstance);
     const int ExitCode = App.Run();
