+++ Src/pch.h
@@ -1,3 +1,6 @@
 #pragma once
 
 #include "DirectXMath.h"
+
+// Global ptr to the main window instance.
+inline class MainWindow* G_MainWindow = nullptr;
